backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
pillow==11.1.0
websocket-client==1.8.0
